Rename atom to atomic
